RecordingTime and BarcodeEditDialog widgets are now in status folder and BarcodeEditDialog is captured in the E2E testing of platebarcode during manual switchover as it was earlier
