+++ pages/status/basic-barcode-editdialog.vue
@@ -5,7 +5,7 @@
 </template>
 
 <script>
-import { BarcodeEditDialog } from "../../../../../dist/mantarray.common";
+import { BarcodeEditDialog } from "../../dist/mantarray.common";
 // import BarcodeEditDialog from "@/components/playback/controls/player/BarcodeEditDialog.vue";
 
 export default {
