+++ pages/status/on-recording-init.vue
@@ -5,12 +5,16 @@
 </template>
 
 <script>
-import { RecordingTime as ComponentToTest } from "../../../../dist/mantarray.common";
 // import ComponentToTest from "@/components/playback/controls/RecordingTime.vue";
+import { RecordingTime as ComponentToTest } from "../../dist/mantarray.common";
 
 export default {
   components: {
     ComponentToTest,
+  },
+  computed: {},
+  created: function () {
+    this.$store.dispatch("playback/start_recording");
   },
 };
 </script>
