+++ pages/status/recording-basic.vue
@@ -1,0 +1,16 @@
+<template>
+  <div>
+    <ComponentToTest></ComponentToTest>
+  </div>
+</template>
+
+<script>
+import { RecordingTime as ComponentToTest } from "../../dist/mantarray.common";
+// import ComponentToTest from "@/components/playback/controls/RecordingTime.vue";
+
+export default {
+  components: {
+    ComponentToTest,
+  },
+};
+</script>
