+++ tests/e2e/screenshots/status/recording-time/basic-init-base.png
