+++ tests/e2e/screenshots/status/recording-time/on-recording-after-time-elapsed-base.png
