+++ tests/e2e/screenshots/status/recording-time/on-recording-init-base.png
