+++ tests/e2e/screenshots/status/recording-time/x-y-offset-base.png
